+++ classes/Login.php
@@ -1,12 +1,13 @@
 <?php
 /**
- * @package    Grav.Plugin.Login
+ * @package    Grav\Plugin\Login
  *
  * @copyright  Copyright (C) 2014 - 2017 RocketTheme, LLC. All rights reserved.
  * @license    MIT License; see LICENSE file for details.
  */
 namespace Grav\Plugin\Login;
 
+use Birke\Rememberme\Cookie;
 use Grav\Common\Config\Config;
 use Grav\Common\Grav;
 use Grav\Common\File\CompiledYamlFile;
@@ -16,6 +17,9 @@ use Grav\Common\User\User;
 use Grav\Common\Uri;
 use Grav\Common\Utils;
 use Grav\Plugin\Email\Utils as EmailUtils;
+use Grav\Plugin\Login\Events\UserLoginEvent;
+use Grav\Plugin\Login\RememberMe\RememberMe;
+use Grav\Plugin\Login\RememberMe\TokenStorage;
 use RocketTheme\Toolbox\Session\Message;
 
 /**
@@ -36,11 +40,14 @@ class Login
     /** @var Session */
     protected $session;
 
-    /** @var User */
-    protected $user;
-
     /** @var Uri */
     protected $uri;
+
+    /** @var RememberMe */
+    protected $rememberMe;
+
+    /** @var RateLimiter[] */
+    protected $rateLimiters = [];
 
     /**
      * Login constructor.
@@ -53,8 +60,108 @@ class Login
         $this->config = $this->grav['config'];
         $this->language = $this->grav['language'];
         $this->session = $this->grav['session'];
-        $this->user = $this->grav['user'];
         $this->uri = $this->grav['uri'];
+    }
+
+    /**
+     * Login user.
+     *
+     * @param array $credentials
+     * @param array $options
+     * @return User
+     */
+    public function login(array $credentials, array $options = [])
+    {
+        $grav = Grav::instance();
+
+        $eventOptions = [
+            'credentials' => $credentials,
+            'options' => $options
+        ];
+
+        // Attempt to authenticate the user.
+        $event = new UserLoginEvent($eventOptions);
+        $grav->fireEvent('onUserLoginAuthenticate', $event);
+
+        if ($event->isSuccess()) {
+
+            // Make sure that event didn't mess up with the user authorization.
+            $user = $event->getUser();
+            $user->authenticated = true;
+            $user->authorized = false;
+
+            // Allow plugins to prevent login after successful authentication.
+            $event = new UserLoginEvent($event->toArray());
+            $grav->fireEvent('onUserLoginAuthorize', $event);
+        }
+
+        if ($event->isSuccess()) {
+            // User has been logged in, let plugins know.
+            $event = new UserLoginEvent($event->toArray());
+            $grav->fireEvent('onUserLogin', $event);
+
+            // Make sure that event didn't mess up with the user authorization.
+            $user = $event->getUser();
+            $user->authenticated = true;
+            $user->authorized = $event->isDelayed();
+
+        } else {
+            // Allow plugins to log errors or do other tasks on failure.
+            $event = new UserLoginEvent($event->toArray());
+            $grav->fireEvent('onUserLoginFailure', $event);
+
+            // Make sure that event didn't mess up with the user authorization.
+            $user = $event->getUser();
+            $user->authenticated = false;
+            $user->authorized = false;
+        }
+
+        $user = $event->getUser();
+        $user->def('language', 'en');
+
+        return $user;
+    }
+
+    /**
+     * Logout user.
+     *
+     * @param array $options
+     * @param User $user
+     * @return User
+     */
+    public function logout(array $options = [], User $user = null)
+    {
+        $grav = Grav::instance();
+
+        $eventOptions = [
+            'user' => $user ?: $grav['user'],
+            'options' => $options
+        ];
+
+        $event = new UserLoginEvent($eventOptions);
+
+        // Logout the user.
+        $grav->fireEvent('onUserLogout', $event);
+
+        $user = $event->getUser();
+        $user->authenticated = false;
+
+        return $user;
+    }
+
+    /**
+     * Get Current logged in user
+     *
+     * @return User
+     */
+    public function getUser()
+    {
+        /** @var User $user */
+        $user = $this->grav['user'];
+        if (!($user->get('access') || $user->get('groups'))) {
+            $user = User::load($user->get('username'));
+        }
+        return $user;
     }
 
     /**
@@ -83,6 +190,29 @@ class Login
         $messages = $this->grav['messages'];
 
         return $messages->fetch($type);
+    }
+
+    /**
+     * Authenticate user.
+     *
+     * @param array $credentials Form fields.
+     * @param array $options
+     *
+     * @return bool
+     */
+    public function authenticate($credentials, $options = ['remember_me' => true])
+    {
+        $user = $this->login($credentials, $options);
+
+        if ($user->authenticated) {
+            $this->setMessage($this->language->translate('PLUGIN_LOGIN.LOGIN_SUCCESSFUL',
+                [$user->language]), 'info');
+
+            $redirect_route = $this->uri->route();
+            $this->grav->redirect($redirect_route);
+        }
+
+        return $user->authenticated;
     }
 
     /**
@@ -126,15 +256,15 @@ class Login
         return $user;
     }
 
-
     /**
      * Handle the email to notify the user account creation to the site admin.
      *
-     * @param $user
+     * @param User $user
      *
      * @return bool True if the action was performed.
+     * @throws \RuntimeException
      */
-    public function sendNotificationEmail($user)
+    public function sendNotificationEmail(User $user)
     {
         if (empty($user->email)) {
             throw new \RuntimeException($this->language->translate('PLUGIN_LOGIN.USER_NEEDS_EMAIL_FIELD'));
@@ -167,11 +297,12 @@ class Login
     /**
      * Handle the email to welcome the new user
      *
-     * @param $user
+     * @param User $user
      *
      * @return bool True if the action was performed.
+     * @throws \RuntimeException
      */
-    public function sendWelcomeEmail($user)
+    public function sendWelcomeEmail(User $user)
     {
         if (empty($user->email)) {
             throw new \RuntimeException($this->language->translate('PLUGIN_LOGIN.USER_NEEDS_EMAIL_FIELD'));
@@ -198,8 +329,9 @@ class Login
      * @param User $user
      *
      * @return bool True if the action was performed.
+     * @throws \RuntimeException
      */
-    public function sendActivationEmail($user)
+    public function sendActivationEmail(User $user)
     {
         if (empty($user->email)) {
             throw new \RuntimeException($this->language->translate('PLUGIN_LOGIN.USER_NEEDS_EMAIL_FIELD'));
@@ -234,13 +366,78 @@ class Login
     }
 
     /**
+     * Gets and sets the RememberMe class
+     *
+     * @param  mixed $var A rememberMe instance to set
+     *
+     * @return RememberMe Returns the current rememberMe instance
+     * @throws \InvalidArgumentException
+     */
+    public function rememberMe($var = null)
+    {
+        if ($var !== null) {
+            $this->rememberMe = $var;
+        }
+
+        if (!$this->rememberMe) {
+            /** @var Config $config */
+            $config = $this->grav['config'];
+
+            // Setup storage for RememberMe cookies
+            $storage = new TokenStorage;
+            $this->rememberMe = new RememberMe($storage);
+            $this->rememberMe->setCookieName($config->get('plugins.login.rememberme.name'));
+            $this->rememberMe->setExpireTime($config->get('plugins.login.rememberme.timeout'));
+
+            // Hardening cookies with user-agent and random salt or
+            // fallback to use system based cache key
+            $server_agent = isset($_SERVER['HTTP_USER_AGENT']) ? $_SERVER['HTTP_USER_AGENT'] : 'unknown';
+            $data = $server_agent . $config->get('security.salt', $this->grav['cache']->getKey());
+            $this->rememberMe->setSalt(hash('sha512', $data));
+
+            // Set cookie with correct base path of Grav install
+            $cookie = new Cookie;
+            $cookie->setPath($this->grav['base_url_relative'] ?: '/');
+            $this->rememberMe->setCookie($cookie);
+        }
+
+        return $this->rememberMe;
+    }
+
+    /**
+     * @param string $context
+     * @param int $maxCount
+     * @param int $interval
+     * @return RateLimiter
+     */
+    public function getRateLimiter($context, $maxCount = null, $interval = null)
+    {
+        if (!isset($this->rateLimiters[$context])) {
+            switch ($context) {
+                case 'login_attempts':
+                    $maxCount = $this->grav['config']->get('plugins.login.max_login_count', 5);
+                    $interval = $this->grav['config']->get('plugins.login.max_login_interval', 10);
+                    break;
+                case 'pw_resets':
+                    $maxCount = $this->grav['config']->get('plugins.login.max_pw_resets_count', 0);
+                    $interval = $this->grav['config']->get('plugins.login.max_pw_resets_interval', 2);
+                    break;
+            }
+            $this->rateLimiters[$context] = new RateLimiter($context, $maxCount, $interval);
+        }
+
+        return $this->rateLimiters[$context];
+    }
+
+    /**
      * Check if user may use password reset functionality.
      *
-     * @param  User $user
-     * @param $field
-     * @param $count
-     * @param $interval
+     * @param User   $user
+     * @param string $field
+     * @param int    $count
+     * @param int    $interval
      * @return bool
+     * @deprecated 3.0 Use $grav['login']->getRateLimiter($context) instead. See Grav\Plugin\Login\RateLimiter class.
      */
     public function isUserRateLimited(User $user, $field, $count, $interval)
     {
@@ -266,15 +463,56 @@ class Login
         return false;
     }
 
+
+    public function isUserAuthorizedForPage($user, $page, $config = null)
+    {
+        $header = $page->header();
+        $rules = isset($header->access) ? (array)$header->access : [];
+
+        if ($config && $config->get('parent_acl')) {
+            // If page has no ACL rules, use its parent's rules
+            if (!$rules) {
+                $parent = $page->parent();
+                while (!$rules and $parent) {
+                    $header = $parent->header();
+                    $rules = isset($header->access) ? (array)$header->access : [];
+                    $parent = $parent->parent();
+                }
+            }
+        }
+
+        // Continue to the page if it has no ACL rules.
+        if (!$rules) {
+            return true;
+        }
+
+        // Continue to the page if user is authorized to access the page.
+        foreach ($rules as $rule => $value) {
+            if (is_array($value)) {
+                foreach ($value as $nested_rule => $nested_value) {
+                    if ($user->authorize($rule . '.' . $nested_rule) == $nested_value) {
+                        return true;
+                    }
+                }
+            } else {
+                if ($user->authorize($rule) == $value) {
+                    return true;
+                }
+            }
+        }
+
+        return false;
+    }
+
     /**
-     * Reset the rate limit counter
+     * Reset the rate limit counter.
      *
-     * @param User $user
-     * @param $field
+     * @param User   $user
+     * @param string $field
+     * @deprecated 3.0 Use $grav['login']->getRateLimiter($context) instead. See Grav\Plugin\Login\RateLimiter class.
      */
     public function resetRateLimit(User $user, $field)
     {
         $user->{$field} = [];
     }
-
 }